backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
pandas
requests
python-dateutil
cachetools
